ies_pipeline: TE support for NSH encap/decap
The tunnel engine now can encap and decap VXLAN-GPE+NSH packets.
For encap, the same VXLAN parameters are required, in addition to
the NSH Service Path ID and the NSH Service Index.

Only NSH MD-type 1 is supported, and only supports encapsulation
of Ethernet frames.

On decap, the hardware expects the packet has the length of the
NSH base, service, and context headers to be present in the same
location as the GENEVE 'Opt Len' field.

Only packets with a destination UDP port of 4790 can be decapsulated
as this is the assigned VXLAN-GPE port.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Hao Zheng <hao.zheng@intel.com>