matchlib: add syslog-like logging system
Previously all messages were logged to stdout or stderr. Some
messages are wrapped in '#ifdef DEBUG' blocks to only be
emitted during debug.

This patch provides a mechanism very similar to syslog, with
support for compiling out debug/log messages which are not
intended to be printed.

It also allows for applications to override the default log
functions to better integrate with applications that prefer
to do their own logging.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: John Fastabend <john.r.fastabend@intel.com>