matchd: enforce correct mask specification
When a table is created, each match requires a mask type
to be specified. Previously there was no check to make sure
the requested mask is supported by the match field. This patch
checks that a requested mask is the same as the specified mask.

When a rule is added for a field containing an exact match, the
mask must have all bits set, hence exact match. Previously there
was no check to ensure the match was valid.  This patch verifies
the match and if invalid, emits an error message and rejects the
rule.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Hao Zheng <hao.zheng@intel.com>