ieslib: rework lport lookup scheme
Prior to IES 4.1.3, there was not a straightforward way to
determine the logical port number for a given PCI address. To
workaround the limitation, the ieslib backend attempted to gather
the information by reading the MAC Address table and comparing the
glort of each entry against the value expected to be assigned
based on empirical testing. If the glort matched the expected
value, the ieslib used the logical port number from the MA table.

As of IES 4.1.3 an API exists to determine the logical port
given the PEP number and VF index.

This commit cleans up the old workaround and uses the new API
in when setting forward_vsi actions and executing the GET_LPORT
command.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: John Fastabend <john.r.fastabend@intel.com>