match: fix invalid free when creating tables
The action name field may or may not be populated by the
backend calls from libmatch.

If the field is not populated and because it is not initialized
to 0 we can get an invalid free and possible segfault. This
patch resolves the issue by initializing the field to 0 and it
is always safe to free a zero address.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Acked-by: Jeff Shaw <jeffrey.b.shaw@intel.com>