match: include string names in rule sets
The backend only records the rule names if the client provides the
strings. The logic here is if the client is not using string names
then it should not expect or need to receive string names on queries.

The match CLI does not provide string names and so they are not
returned by the backend. This makes pretty printing a bit clumsy.
To clean this up we have two options either the match CLI could keep
its own database of UID to string lookups _or_ we can simply include
them in the rule commands sent to matchd.

This patch does the simplest thing and adds them to the message sent
to build rules.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Acked-by: Jeff Shaw <jeffrey.b.shaw@intel.com>