ieslib: fix setting port speed
Previously the ies backend was attempting to set the port
speed using the FM_PORT_SPEED attribute, which is read-only.

Actually, the port speed is driven by the Ethernet mode so
we need to instead use the FM_PORT_ETHERNET_INTERFACE_MODE
attribute.

Fiber interfaces can be set easily. Many of the copper modes
are required to be auto-negotiated which is currently not
supported.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Sushma Sitaram <sushma.sitaram@intel.com>
Acked-by: Hao Zheng <hao.zheng@intel.com>