doc: doxygen build infrastructure
The API documentation will be generated by doxygen if the build
environment is configured to do so.  Enable doc building with:

 $ ./configure --enable-doc
 $ make

The API documentation is generated in HTML format in doc/api.
Point your browser to doc/api/index.html to preview the generated
output. Currently no files are processed as documentation is WIP.

Once a header file has been documented, add it to the FILE_PATTERNS
list in doxyfile.in and re-configure and build.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Hao Zheng <hao.zheng@intel.com>