match: allow port 0
Previously it was assumed that port_id == 0 meant the end
of a port list.  This is not the case when ports actually
have an id of 0 as is the case with the IES cpu port.

This commit defines a NET_MAT_PORT_ID_UNSPEC which is used
to indicate the end of a port list.

Fixes: ONP-2139

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Sushma Sitaram <sushma.sitaram@intel.com>