matchd_lib: fix ONP-2083 issue, match hangs with wrong netlink family number
Previously,  in matchd_init, matchd did not verify if the input
netlink family id was valid.

In matchd_rx_process, nlh->nlmsg_type which stores the family id
was checked if equal to NLMSG_ERROR. NLMSG_ERROR has a value 0x2.
This caused match to hang when run with family id  = 2. This patch
checks if nlh->nlmsg_type equals family id and resolves the issue.
It returns an error in cases where nlmsg_type is not equal to family id.

Signed-off-by: Sushma Sitaram <sushma.sitaram@intel.com>
Acked-by: Indumathi Madhavan <indumathi.madhavan@intel.com>