ieslib: disable ies-tagging
In switch_init() we disable ies-tagging for all the host-interfaces.
With ies-tagging enabled, the switch inserts some metadata at the
beginning of the frame before the Ethernet header. Unless the
frames are received and processed with a raw socket, the network
stack will not know how to disposition the packet.

The IES software enables this by default for the interface it
believes is the "CPU" interface.  For normal networking to function
we need to have it disabled.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Hao Zheng <hao.zheng@intel.com>